integrate UI with cubit
